test: consolidate IRC socket fixture lifetimes
